requirements.txt
anthropic==0.8.0
gitpython==3.1.42
pygithub==2.1.1
python-dotenv==1.0.1
requests==2.31.0
pydantic==2.6.0
typing-extensions==4.9.0
